// bitlo-angular-case/src/app/shared/models/markets.model.ts
export interface MarketData {
    ask: string;
    bid: string;
    change24h: string;
    change24hPercent: string;
    currentQuote: string;
    highestQuote24h: string;
    lowestQuote24h: string;
    marketCode: string;
    notionalVolume24h: string;
    volume24h: string;
    weightedAverage24h: string;
}